// MainWebSite/src/app/Services/credits/models/get-payment-list-credit-result.ts
export class GetPaymentListCreditResult {
    movementDate: string;
    movementHour: string;
    gloss: string;
    amount: number;
    branchOffice: string;
    agency: string;
    userName: string;
    operationNumber: number;
    originalBalance: number;
    principalBalance: number;
    endBalance: number;
    accountNumber: string;
    currency: string;
}